+++ src/util/ShowProgressEvent.js
@@ -1,0 +1,21 @@
+'use strict'
+
+class ShowProgressEvent {
+  constructor (progressBar) {
+    if (progressBar) {
+      return (event) => {
+        if (event.lengthComputable) {
+          progressBar.style.display = ''
+          const percentComplete = parseInt((event.loaded / event.total) * 100)
+          progressBar.value = percentComplete
+          if (progressBar.value === 100) {
+            progressBar.style.display = 'none'
+          }
+        }
+      }
+    }
+    return () => {}
+  }
+}
+
+module.exports = ShowProgressEvent
